rustc_utils::mir::location_or_arg::index

Struct LocationOrArgIndex

pub struct LocationOrArgIndex { /* private fields */ }

Available on crate feature indexical only.

Implementations

impl LocationOrArgIndex

pub const MAX_INDEX: usize = 4_294_967_295usize

If Self::CHECKS_MAX_INDEX is true, we’ll assert if trying to produce a value larger than this in any of the ctors that don’t have unchecked in their name.

pub const CHECKS_MAX_INDEX: bool = true

Does this index type assert if asked to construct an index larger than MAX_INDEX?

pub fn new(value: usize) -> Self

Construct this index type from a usize. Alias for from_usize.

pub fn from_raw(value: u32) -> Self

Construct this index type from the wrapped integer type.

pub fn from_foreign<F: Idx>(value: F) -> Self

Construct this index type from one in a different domain

pub const fn from_usize_unchecked(value: usize) -> Self

Construct from a usize without any checks.

pub const fn from_raw_unchecked(raw: u32) -> Self

Construct from the underlying type without any checks.

pub fn from_usize(value: usize) -> Self

Construct this index type from a usize.

pub const fn index(self) -> usize

Get the wrapped index as a usize.

pub const fn raw(self) -> u32

Get the wrapped index.

pub fn check_index(v: usize)

Asserts v <= Self::MAX_INDEX unless Self::CHECKS_MAX_INDEX is false.