Module rustc_utils::mir::place

Utilities for Place.

Structs

PlaceCollector

A MIR Visitor which collects all Places that appear in the visited object.

SimplifyMir

MIR pass to remove instructions not important for Flowistry.

Constants

UNKNOWN_REGION

Used to describe aliases of owned and raw pointers.

Traits

PlaceExt

Extension trait for Place.