rustc_utils

Module test_utils

Available on crate feature test only.

Running rustc and Flowistry in tests.

Structs

CompileBuilder

Programmatically build a rustc compilation session

CompileResult

State during the rust compilation. Most of the time you only care about self.tcx, but this wrapper provides additional convenience methods for getting, e.g. the body of the configured entrypoint.

PlaceBuilder

Placer

StringLoader

Constants

DUMMY_BYTE_RANGE

DUMMY_CHAR_RANGE

DUMMY_FILE

DUMMY_FILE_NAME

Functions

color_ranges

compare_ranges

compare_sets

compile_body

Convenience alias for CompileBuilder::new(input).compile(...) if the callback is going to use CompileResult::as_body.

fmt_ranges

parse_ranges

Type Aliases

RangeMap