Trait rustc_utils::hir::ty::TyExt

pub trait TyExt<'tcx> {
    type AllRegionsIter<'a>: Iterator<Item = Region<'tcx>>
       where Self: 'a;

    // Required methods
    fn inner_regions(&self) -> Self::AllRegionsIter<'_>;
    fn does_implement_trait(
        &self,
        tcx: TyCtxt<'tcx>,
        param_env: ParamEnv<'tcx>,
        trait_def_id: DefId
    ) -> bool;
    fn is_copyable(&self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>) -> bool;
}

Extension trait for Ty.

Required Associated Types

type AllRegionsIter<'a>: Iterator<Item = Region<'tcx>> where Self: 'a

Required Methods

fn inner_regions(&self) -> Self::AllRegionsIter<'_>

Returns an iterator over the regions appearing within a type.

fn does_implement_trait( &self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>, trait_def_id: DefId ) -> bool

Returns true if a type implements a given trait.

fn is_copyable(&self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>) -> bool

Returns true if a type implements Copy.

Object Safety

This trait is not object safe.